+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## Unreleased
 
+### Changed
+
+* Switch to rollup for package compilation.
+* Move `DraftailEditor` from default export of draftail to named export (`import { DraftailEditor } from 'draftail';`).
+
 ## [[v0.10.0]](https://github.com/springload/draftail/releases/tag/v0.10.0)
 
 ### Added